refactor(e2e): extract SingleNodeTestContext base + shared multi-node setup presets
Carve the prod-sequencer single-node lifecycle out of MultiNodeTestContext into a
new SingleNodeTestContext parent (environment, node spawning, prover lifecycle, and
the epoch/checkpoint/proof-window/reorg waiters), leaving MultiNodeTestContext with
only the validator-node spawning and committee-convergence helpers. Set inboxLag: 2
as a base default (the intended value when pipelining). Add shared presets/helpers:
buildMockGossipValidators, MOCK_GOSSIP_MULTI_VALIDATOR_OPTS, FAST_REORG_TIMING,
defaultSlashingPenalties/withOnlyOffense. Rename multi-node/e2e_slashing -> slashing
(git mv + describe titles + .test_patterns.yml + bootstrap.sh discovery). All 29
tests still import MultiNodeTestContext and pass unchanged.
